+++ Enums/SyncAction.cs
@@ -1,9 +1,9 @@
 ﻿namespace FileSync.Enums
 {
-    public enum SyncAction
-    {
-        Create,
-        Update,
-        Delete
-    }
+  public enum SyncAction
+  {
+    Create,
+    Update,
+    Delete
+  }
 }
