+++ Program.cs
@@ -1,25 +1,35 @@
-﻿namespace FileSync
+﻿using System;
+using System.Windows.Forms;
+using FileSync.Views;
+using FileSync.Logging;
+using FileSync.Presenter;
+
+namespace FileSync
 {
-    using System;
-    using System.Windows.Forms;
-    using FileSync.Views;
-    using FileSync.Logging;
-    using FileSync.Presenter;
-
-    internal static class Program
+  internal static class Program
+  {
+    [STAThread]
+    private static void Main()
     {
-        [STAThread]
-        private static void Main()
+      Application.EnableVisualStyles();
+      Application.SetCompatibleTextRenderingDefault(false);
+
+      var view = new SyncView();
+      var logManagerFactory = new Func<bool, ILogManager>(useJson =>
+      {
+        if (useJson)
         {
-            Application.EnableVisualStyles();
-            Application.SetCompatibleTextRenderingDefault(false);
-
-            var view = new SyncView();
-            var logManagerFactory = new Func<bool, ILogManager>(useJson =>
-                useJson ? new JsonLogManager("sync_log.json") : new XmlLogManager("sync_log.xml"));
-            var presenter = new SyncPresenter(view, logManagerFactory);
-
-            Application.Run(view);
+          return new JsonLogManager("sync_log.json");
         }
+        else
+        {
+          return new XmlLogManager("sync_log.xml");
+        }
+      });
+      
+      var presenter = new SyncPresenter(view, logManagerFactory);
+
+      Application.Run(view);
     }
+  }
 }
