+++ Services/FileComparer.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Security.Cryptography;
+using FileSync.Model;
 
 public class FileComparer
 {
